LPS-171473 Renaming to factory and moving to the operation package
