+++ modules/apps/headless/headless-builder/headless-builder-api/bnd.bnd
@@ -5,5 +5,4 @@ Export-Package:\
 	com.liferay.headless.builder.model,\
 	com.liferay.headless.builder.operation,\
 	com.liferay.headless.builder.operation.handler,\
-	com.liferay.headless.builder.operation.response,\
-	com.liferay.headless.builder.provider
+	com.liferay.headless.builder.operation.response
